Extract common elements into ConfigSchemaProcessor.kt
