+++ prime-router/src/main/kotlin/fhirengine/translation/hl7/FhirToHl7Converter.kt
@@ -3,17 +3,14 @@ package gov.cdc.prime.router.fhirengine.translation.hl7
 import ca.uhn.hl7v2.HL7Exception
 import ca.uhn.hl7v2.model.Message
 import ca.uhn.hl7v2.util.Terser
-import gov.cdc.prime.router.fhirengine.translation.hl7.schema.ConfigSchemaElement
 import gov.cdc.prime.router.fhirengine.translation.hl7.schema.converter.ConverterSchema
 import gov.cdc.prime.router.fhirengine.translation.hl7.schema.converter.ConverterSchemaElement
 import gov.cdc.prime.router.fhirengine.translation.hl7.schema.converter.converterSchemaFromFile
 import gov.cdc.prime.router.fhirengine.translation.hl7.utils.ConstantSubstitutor
 import gov.cdc.prime.router.fhirengine.translation.hl7.utils.CustomContext
-import gov.cdc.prime.router.fhirengine.translation.hl7.utils.FhirPathUtils
 import gov.cdc.prime.router.fhirengine.translation.hl7.utils.HL7Utils
 import org.apache.commons.io.FilenameUtils
 import org.apache.logging.log4j.Level
-import org.apache.logging.log4j.kotlin.Logging
 import org.hl7.fhir.r4.model.Base
 import org.hl7.fhir.r4.model.Bundle
 
@@ -31,7 +28,7 @@ class FhirToHl7Converter(
     private var terser: Terser? = null,
     // the constant substitutor is not thread save, so we need one instance per converter instead of using a shared copy
     private val constantSubstitutor: ConstantSubstitutor = ConstantSubstitutor()
-) : Logging {
+) : ConfigSchemaProcessor() {
     /**
      * Convert a FHIR bundle to an HL7 message using the [schema] in the [schemaFolder] location to perform the conversion.
      * The converter will error out if [strict] is set to true and there is an error during the conversion.  if [strict]
@@ -131,7 +128,7 @@ class FhirToHl7Converter(
         var debugMsg = "Processed element name: ${element.name}, required: ${element.required}, "
 
         // First we need to resolve a resource value if available.
-        val focusResources = getFocusResources(element, bundle, focusResource, elementContext)
+        val focusResources = getFocusResources(element.resource, bundle, focusResource, elementContext)
         if (focusResources.isEmpty() && element.required == true) {
             // There are no sources to parse, but the element was required
             throw RequiredElementException(element)
@@ -163,7 +160,7 @@ class FhirToHl7Converter(
 
                     // A value
                     element.value.isNotEmpty() && element.hl7Spec.isNotEmpty() -> {
-                        val value = getValue(element, bundle, singleFocusResource, elementContext)
+                        val value = getValueAsString(element, bundle, singleFocusResource, elementContext)
                         setHl7Value(element, value, context)
                         debugMsg += "condition: true, resourceType: ${singleFocusResource.fhirType()}, " +
                             "value: $value, hl7Spec: ${element.hl7Spec}"
@@ -182,88 +179,6 @@ class FhirToHl7Converter(
         // Only log for elements that require values
         if (element.schemaRef == null) logger.log(logLevel, debugMsg)
         logger.trace("End processing of element ${element.name}.")
-    }
-
-    /**
-     * Get the first valid string from the list of values specified in the schema for a given [element] using
-     * [bundle] and [context] starting at the [focusResource].
-     * @return the value for the element or an empty string if no value found
-     */
-    internal fun getValue(
-        element: ConfigSchemaElement,
-        bundle: Bundle,
-        focusResource: Base,
-        context: CustomContext
-    ): String {
-        var retVal = ""
-        run findValue@{
-            element.value.forEach {
-                val value = if (it.isBlank()) ""
-                else try {
-                    FhirPathUtils.evaluateString(context, focusResource, bundle, it)
-                } catch (e: SchemaException) {
-                    logger.error("Error while getting value for element ${element.name}", e)
-                    ""
-                }
-                logger.trace("Evaluated value expression '$it' to '$value'")
-                if (value.isNotBlank()) {
-                    retVal = value
-                    return@findValue
-                }
-            }
-        }
-
-        // when valueSet is available, use the matching value else just pass the value as is
-        // does a lowerCase comparison
-        if (element.valueSet.isNotEmpty()) {
-            val lowerSet = element.valueSet.mapKeys { it.key.lowercase() }
-            retVal = lowerSet.getOrDefault(retVal.lowercase(), retVal)
-        }
-        return retVal
-    }
-
-    /**
-     * Determine the focus resource for an [element] using [bundle] and the [previousFocusResource].
-     * @return a list of focus resources containing at least one resource.  Multiple resources are returned for collections
-     */
-    internal fun getFocusResources(
-        element: ConfigSchemaElement,
-        bundle: Bundle,
-        previousFocusResource: Base,
-        context: CustomContext
-    ): List<Base> {
-        val resourceList = if (element.resource == null) {
-            listOf(previousFocusResource)
-        } else {
-            val evaluatedResource = FhirPathUtils
-                .evaluate(context, previousFocusResource, bundle, element.resource!!)
-            evaluatedResource
-        }
-
-        return resourceList
-    }
-
-    /**
-     * Test if an [element] can be evaluated based on the [element]'s condition.  Use the [bundle] and [focusResource] * to evaluate the condition expression.
-     * @return true if the condition expression evaluates to a boolean or if the condition expression is empty, false otherwise
-     */
-    internal fun canEvaluate(
-        element: ConfigSchemaElement,
-        bundle: Bundle,
-        focusResource: Base,
-        context: CustomContext
-    ): Boolean {
-        return element.condition?.let {
-            try {
-                FhirPathUtils.evaluateCondition(context, focusResource, bundle, it)
-            } catch (e: SchemaException) {
-                logger.warn(
-                    "Condition for element ${element.name} did not evaluate to a boolean type, " +
-                        "so the condition failed."
-                )
-                false
-            }
-        } ?: true
     }
 
     /**
