+++ prime-router/src/main/kotlin/fhirengine/translation/hl7/FhirTransformer.kt
@@ -1,6 +1,5 @@
 package gov.cdc.prime.router.fhirengine.translation.hl7
 
-import gov.cdc.prime.router.fhirengine.translation.hl7.schema.ConfigSchemaElement
 import gov.cdc.prime.router.fhirengine.translation.hl7.schema.fhirTransform.FHIRTransformSchemaElement
 import gov.cdc.prime.router.fhirengine.translation.hl7.schema.fhirTransform.FhirTransformSchema
 import gov.cdc.prime.router.fhirengine.translation.hl7.schema.fhirTransform.fhirTransformSchemaFromFile
@@ -10,7 +9,6 @@ import gov.cdc.prime.router.fhirengine.translation.hl7.utils.FhirBundleUtils
 import gov.cdc.prime.router.fhirengine.translation.hl7.utils.FhirPathUtils
 import org.apache.commons.io.FilenameUtils
 import org.apache.logging.log4j.Level
-import org.apache.logging.log4j.kotlin.Logging
 import org.hl7.fhir.r4.model.Base
 import org.hl7.fhir.r4.model.Bundle
 import org.hl7.fhir.r4.model.Extension
@@ -27,7 +25,7 @@ class FhirTransformer(
     private val strict: Boolean = false,
     // the constant substitutor is not thread safe, so we need one instance per converter instead of using a shared copy
     private val constantSubstitutor: ConstantSubstitutor = ConstantSubstitutor()
-) : Logging {
+) : ConfigSchemaProcessor() {
     /**
      * Transform a FHIR bundle based on the [schema] in the [schemaFolder] location.
      * The transformer will error out if [strict] is set to true and there is an error during the translation. If [strict]
@@ -172,91 +170,6 @@ class FhirTransformer(
         // Only log for elements that require values
         if (element.schemaRef == null) logger.log(logLevel, debugMsg)
         logger.trace("End processing of element ${element.name}.")
-    }
-
-    /**
-     * Get the first valid string from the list of values specified in the schema for a given [element] using
-     * [bundle] and [context] starting at the [focusResource].
-     * @return the value for the element or an empty string if no value found
-     */
-    internal fun getValue(
-        element: ConfigSchemaElement,
-        bundle: Bundle,
-        focusResource: Base,
-        context: CustomContext
-    ): Base? {
-        var retVal: Base? = null
-        run findValue@{
-            element.value.forEach {
-                val value = if (it.isBlank()) emptyList<Base>()
-                else try {
-                    FhirPathUtils.evaluate(context, focusResource, bundle, it)
-                } catch (e: SchemaException) {
-                    logger.error("Error while getting value for element ${element.name}", e)
-                    emptyList<Base>()
-                }
-                logger.trace("Evaluated value expression '$it' to '$value'")
-                if (value.isNotEmpty()) {
-                    retVal = value[0]
-                    return@findValue
-                }
-            }
-        }
-
-        // when valueSet is available, use the matching value else just pass the value as is
-        // does a lowerCase comparison
-        if (element.valueSet.isNotEmpty()) {
-            val lowerSet = element.valueSet.mapKeys { it.key.lowercase() }
-            val valStr =
-                lowerSet.getOrDefault(retVal?.primitiveValue()?.lowercase() ?: "", retVal?.primitiveValue() ?: "")
-            retVal = FhirPathUtils.evaluate(context, focusResource, bundle, valStr)[0]
-        }
-        return retVal
-    }
-
-    /**
-     * Determine the focus resource from [resourceStr] using [bundle] and the [previousFocusResource].
-     * @return a list of focus resources containing at least one resource.  Multiple resources are returned for collections
-     */
-    internal fun getFocusResources(
-        resourceStr: String?,
-        bundle: Bundle,
-        previousFocusResource: Base,
-        context: CustomContext
-    ): List<Base> {
-        val resourceList = if (resourceStr == null || resourceStr == "") {
-            listOf(previousFocusResource)
-        } else {
-            val evaluatedResource = FhirPathUtils
-                .evaluate(context, previousFocusResource, bundle, resourceStr)
-            evaluatedResource
-        }
-
-        return resourceList
-    }
-
-    /**
-     * Test if an [element] can be evaluated based on the [element]'s condition.  Use the [bundle] and [focusResource]
-     * to evaluate the condition expression.
-     * @return true if the condition expression evaluates to a boolean or if the condition expression is empty, false otherwise
-     */
-    internal fun canEvaluate(
-        element: ConfigSchemaElement,
-        bundle: Bundle,
-        focusResource: Base,
-        context: CustomContext
-    ): Boolean {
-        return element.condition?.let {
-            try {
-                FhirPathUtils.evaluateCondition(context, focusResource, bundle, it)
-            } catch (e: SchemaException) {
-                logger.warn(
-                    "Condition for element ${element.name} did not evaluate to a boolean type, " +
-                        "so the condition failed."
-                )
-                false
-            }
-        } ?: true
     }
 
     /**
